V and VM are private for controller.
I think that they should not be anyhow visible outside.
If anything happens with them, that's governed by controller.
It binds the world changes to MV. It contains methods if other
controllers need to interact directly (not be events).
So it's a kind of widget with one public API. Sounds familiar?